实操/基于softmax线性回归的性别识别系统/requirements.txt
torch~=2.1.0.dev20230408
pillow~=9.5.0
torchvision~=0.16.0.dev20230408
tqdm~=4.65.0
opencv-python~=4.7.0.72